Module syntect::easy

API wrappers for common use cases like highlighting strings and files without caring about intermediate semantic representation and caching.

Structs

HighlightFile
Convenience struct containing everything you need to highlight a file. Use the reader to get the lines of the file and the highlight_lines to highlight them. See the new method docs for more information.

HighlightLines
Simple way to go directly from lines of text to coloured tokens.

ScopeRegionIterator
Iterator over the regions of a line which a given the operation from the parser applies.