Module syntect::highlighting::selector

Structs

ScopeSelector
A single selector consisting of a stack to match and a possible stack to exclude from being matched. You probably want ScopeSelectors which is this but with union support.

ScopeSelectors
A selector set that matches anything matched by any of its component selectors. See The TextMate Docs for how these work.

Constants

_IMPL_DESERIALIZE_FOR_ScopeSelector

_IMPL_DESERIALIZE_FOR_ScopeSelectors

_IMPL_SERIALIZE_FOR_ScopeSelector

_IMPL_SERIALIZE_FOR_ScopeSelectors