Struct syntect::highlighting::HighlightState

pub struct HighlightState {
    pub path: ScopeStack,
    // some fields omitted
}

Keeps a stack of scopes and styles as state between highlighting different lines. If you are highlighting an entire file you create one of these at the start and use it all the way to the end.

Caching

One reason this is exposed is that since it implements Clone you can actually cache these (probably along with a ParseState) and only re-start highlighting from the point of a change. You could also do something fancy like only highlight a bit past the end of a user's screen and resume highlighting when they scroll down on large files.

Alternatively you can save space by caching only the path field of this struct then re-create the HighlightState when needed by passing that stack as the initial_stack parameter to the new method. This takes less space but a small amount of time to re-create the style stack.

Note: Caching is for advanced users who have tons of time to maximize performance or want to do so eventually. It is not recommended that you try caching the first time you implement highlighting.

Fields

path

Methods

impl HighlightState

fn new(highlighter: &Highlighter, initial_stack: ScopeStack) -> HighlightState

Note that the Highlighter is not stored, it is used to construct the initial stack of styles. Most of the time you'll want to pass an empty stack as initial_stack but see the docs for HighlightState for discussion of advanced caching use cases.