Struct syntect::highlighting::StyleModifier

pub struct StyleModifier {
    pub foreground: Option<Color>,
    pub background: Option<Color>,
    pub font_style: Option<FontStyle>,
}

A change to a Style applied incrementally by a theme rule.

Fields
foreground: Option<Color>

Foreground color.

background: Option<Color>

Background color.

font_style: Option<FontStyle>

Style of the font.

Methods

impl StyleModifier

pub fn apply(&self, other: StyleModifier) -> StyleModifier

Applies the other modifier to this one, creating a new modifier. Values in other are preferred.