Struct syntect::highlighting::ThemeSet

pub struct ThemeSet {
    pub themes: BTreeMap<String, Theme>,
}

Fields

themes

Methods

impl ThemeSet

A set of themes, includes convenient methods for loading and discovering themes.

fn discover_theme_paths<P: AsRef<Path>>(folder: P) -> Result<Vec<PathBuf>, LoadingError>

Returns all the themes found in a folder, good for enumerating before loading one with get_theme

fn get_theme<P: AsRef<Path>>(path: P) -> Result<Theme, LoadingError>

Loads a theme given a path to a .tmTheme file

fn load_from_folder<P: AsRef<Path>>(folder: P) -> Result<ThemeSet, LoadingError>

Loads all the themes in a folder

impl ThemeSet

fn load_defaults() -> ThemeSet

Loads the set of default themes Currently includes Solarized light/dark, Base16 ocean/mocha/eighties and InspiredGithub