Module syntect::highlighting::style

Structs

Color
RGBA colour, these numbers come directly from the theme so for now you might have to do your own colour space conversion if you are outputting a different colour space from the theme. This can be a problem because some Sublime themes use sRGB and some don't. This is specified in an attribute syntect doesn't parse yet.

FontStyle
This can be a combination of BOLD, UNDERLINE and ITALIC

Style
The foreground, background and font style

StyleModifier
A change to a Style applied incrementally by a theme rule.

Constants

_IMPL_DESERIALIZE_FOR_Color

_IMPL_DESERIALIZE_FOR_FontStyle

_IMPL_DESERIALIZE_FOR_Style

_IMPL_DESERIALIZE_FOR_StyleModifier

_IMPL_SERIALIZE_FOR_Color

_IMPL_SERIALIZE_FOR_FontStyle

_IMPL_SERIALIZE_FOR_Style

_IMPL_SERIALIZE_FOR_StyleModifier