Enum syntect::parsing::ParseSyntaxError

pub enum ParseSyntaxError {
    InvalidYaml(ScanError),
    EmptyFile,
    MissingMandatoryKey(&'static str),
    RegexCompileError(Error),
    InvalidScope(ParseScopeError),
    BadFileRef,
    MainMissing,
    TypeMismatch,
}

Variants
InvalidYaml(ScanError)

Invalid YAML file syntax, or at least something yaml_rust can't handle

EmptyFile

The file must contain at least on YAML document

MissingMandatoryKey(&'static str)

Some keys are required for something to be a valid .sublime-syntax

RegexCompileError(Error)

Invalid regex

InvalidScope(ParseScopeError)

A scope that syntect's scope implementation can't handle

BadFileRef

A reference to another file that is invalid

MainMissing

Syntaxes must have a context named "main"

TypeMismatch

Some part of the YAML file is the wrong type (e.g a string but should be a list) Sorry this doesn't give you any way to narrow down where this is. Maybe use Sublime Text to figure it out.