Enum syntect::parsing::ScopeStackOp

pub enum ScopeStackOp {
    Push(Scope),
    Pop(usize),
    Clear(ClearAmount),
    Restore,
    Noop,
}

A change to a scope stack. Generally Noop is only used internally and you don't have to worry about ever getting one back from a public function. Use ScopeStack#apply to apply this change.

Variants
Push(Scope)Pop(usize)Clear(ClearAmount)

used for the clear_scopes feature

Restore

restores cleared scopes

Noop