Module syntect::parsing::syntax_definition

Data structures for representing syntax definitions

Everything here is public becaues I want this library to be useful in super integrated cases like text editors and I have no idea what kind of monkeying you might want to do with the data. Perhaps parsing your own syntax format into this data structure?

Structs

Context

ContextId

An opaque ID for a Context.

MatchIter

Used to iterate over all the match patterns in a context

MatchPattern

SyntaxDefinition

The main data structure representing a syntax definition loaded from a .sublime-syntax file

Enums

ContextReference

MatchOperation

Pattern

Functions

context_iter

Returns an iterator over all the match patterns in this context.

Type Aliases

CaptureMapping