Function syntect::parsing::syntax_definition::context_iter

pub fn context_iter(ctx: ContextPtr) -> MatchIter

Returns an iterator over all the match patterns in this context. It recursively follows include directives. Can only be run on contexts that have already been linked up.