Struct syntect::parsing::syntax_definition::Context

pub struct Context {
    pub meta_scope: Vec<Scope>,
    pub meta_content_scope: Vec<Scope>,
    pub meta_include_prototype: bool,
    pub prototype: Option<ContextPtr>,
    pub uses_backrefs: bool,
    pub patterns: Vec<Pattern>,
}

Fields
meta_scope: Vec<Scope> meta_content_scope: Vec<Scope> meta_include_prototype: bool

This being set false in the syntax file implies this field being set false, but it can also be set falso for contexts that don't include the prototype for other reasons

prototype: Option<ContextPtr>

This is filled in by the linker at link time for contexts that have meta_include_prototype==true and are not included from the prototype.

uses_backrefs: bool patterns: Vec<Pattern>
Methods

impl Context

fn match_at(&self, index: usize) -> &MatchPattern

Returns the match pattern at an index, panics if the thing isn't a match pattern

fn match_at_mut(&mut self, index: usize) -> &mut MatchPattern

Returns a mutable reference, otherwise like match_at