Struct syntect::parsing::syntax_definition::SyntaxDefinition

pub struct SyntaxDefinition {
    pub name: String,
    pub file_extensions: Vec<String>,
    pub scope: Scope,
    pub first_line_match: Option<String>,
    pub hidden: bool,
    pub prototype: Option<ContextPtr>,
    pub variables: HashMap<String, String>,
    pub contexts: HashMap<String, ContextPtr>,
}

The main data structure representing a syntax definition loaded from a .sublime-syntax file. You'll probably only need these as references to be passed around to parsing code.

Some useful public fields are the name field which is a human readable name to display in syntax lists, and the hidden field which means hide this syntax from any lists because it is for internal use.

Fields

name

file_extensions

scope

first_line_match

hidden

prototype

Filled in at link time to avoid serializing it multiple times

variables

contexts

Methods

impl SyntaxDefinition

fn load_from_str(s: &str, lines_include_newline: bool) -> Result<SyntaxDefinition, ParseSyntaxError>

In case you want to create your own SyntaxDefinition's in memory from strings. Generally you should use a SyntaxSet