syntect
Module easy

API wrappers for common use cases like highlighting strings and files without caring about intermediate semantic representation and caching.

Structs

HighlightFile

Convenience struct containing everything you need to highlight a file

HighlightLines

Simple way to go directly from lines of text to colored tokens.

ScopeRangeIterator

Iterator over the ranges of a line which a given the operation from the parser applies.

ScopeRegionIterator

A convenience wrapper over ScopeRangeIterator to return &strs directly.