Module syntect::highlighting

Everything having to do with turning parsed text into styled text.

You might want to check out Theme for its handy text-editor related settings like selection color, ThemeSet for loading themes, as well as things starting with Highlight for how to highlight text.

Structs

Color

RGBA color, directly from the theme

FontStyle

The color-independent styling of a font - i.e. bold, italicized, and/or underlined

HighlightIterator

Highlights a line of parsed code given a HighlightState and line of changes from the parser.

HighlightState

Keeps a stack of scopes and styles as state between highlighting different lines.

Highlighter

Basically a wrapper around a Theme preparing it to be used for highlighting.

RangedHighlightIterator

Highlights a line of parsed code given a HighlightState and line of changes from the parser.

ScopeSelector

A single selector consisting of a stack to match and a possible stack to exclude from being matched.

ScopeSelectors

A selector set that matches anything matched by any of its component selectors.

ScoredStyle

Style

Foreground and background colors, with font style

StyleModifier

A change to a Style applied incrementally by a theme rule

Theme

A theme parsed from a .tmTheme file.

ThemeItem

A component of a theme meant to highlight a specific thing (e.g string literals) in a certain way.

ThemeSet

ThemeSettings

Properties for styling the UI of a text editor

Enums

UnderlineOption