syntect::highlighting

Struct ThemeSet

pub struct ThemeSet {
    pub themes: BTreeMap<String, Theme>,
}

Fields
themes: BTreeMap<String, Theme>
Implementations

impl ThemeSet

pub fn load_defaults() -> ThemeSet

Loads the set of default themes Currently includes (these are the keys for the map):

base16-ocean.dark,base16-eighties.dark,base16-mocha.dark,base16-ocean.light

InspiredGitHub from here

Solarized (dark) and Solarized (light)

impl ThemeSet

A set of themes, includes convenient methods for loading and discovering themes.

pub fn new() -> ThemeSet

Creates an empty set

pub fn discover_theme_paths<P: AsRef<Path>>( folder: P, ) -> Result<Vec<PathBuf>, LoadingError>

Returns all the themes found in a folder

This is good for enumerating before loading one with get_theme

pub fn get_theme<P: AsRef<Path>>(path: P) -> Result<Theme, LoadingError>

Loads a theme given a path to a .tmTheme file

pub fn load_from_reader<R: BufRead + Seek>( r: &mut R, ) -> Result<Theme, LoadingError>

Loads a theme given a readable stream

pub fn load_from_folder<P: AsRef<Path>>( folder: P, ) -> Result<ThemeSet, LoadingError>

Generate a ThemeSet from all themes in a folder

pub fn add_from_folder<P: AsRef<Path>>( &mut self, folder: P, ) -> Result<(), LoadingError>

Load all the themes in the folder into this ThemeSet