Struct syntect::highlighting::ThemeSet

pub struct ThemeSet {
    pub themes: BTreeMap<String, Theme>,
}

Fields
themes: BTreeMap<String, Theme>
Implementations

impl ThemeSet

pub fn load_defaults() -> ThemeSet

Loads the set of default themes Currently includes (these are the keys for the map):

base16-ocean.dark,base16-eighties.dark,base16-mocha.dark,base16-ocean.light

InspiredGitHub from here

Solarized (dark) and Solarized (light)

impl ThemeSet

A set of themes, includes convenient methods for loading and discovering themes.

pub fn new() -> ThemeSet

Creates an empty set

pub fn discover_theme_paths<P: AsRef<Path>>( folder: P ) -> Result<Vec<PathBuf>, LoadingError>

Returns all the themes found in a folder

This is good for enumerating before loading one with get_theme