Struct syntect::highlighting::ThemeSettings

pub struct ThemeSettings {
    pub foreground: Option<Color>,
    pub background: Option<Color>,
    pub caret: Option<Color>,
    pub line_highlight: Option<Color>,
    pub bracket_contents_foreground: Option<Color>,
    pub bracket_contents_options: Option<UnderlineOption>,
    pub brackets_foreground: Option<Color>,
    pub brackets_background: Option<Color>,
    pub brackets_options: Option<UnderlineOption>,
    pub tags_foreground: Option<Color>,
    pub tags_options: Option<UnderlineOption>,
    pub find_highlight: Option<Color>,
    pub find_highlight_foreground: Option<Color>,
    pub gutter: Option<Color>,
    pub gutter_foreground: Option<Color>,
    pub selection: Option<Color>,
    pub selection_background: Option<Color>,
    pub selection_border: Option<Color>,
    pub inactive_selection: Option<Color>,
    pub guide: Option<Color>,
    pub active_guide: Option<Color>,
    pub stack_guide: Option<Color>,
    pub highlight: Option<Color>,
    pub highlight_foreground: Option<Color>,
}

Various properties meant to be used to style a text editor. Basically all the styles that aren't directly applied to text like selection colour. Use this to make your editor UI match the highlighted text.

Fields

foreground

Foreground color for the view.

background

Backgound color of the view.

caret

Color of the caret.

line_highlight

Color of the line the caret is in. Only used when the higlight_line setting is set to true.

bracket_contents_foreground

Color of bracketed sections of text when the caret is in a bracketed section. Only applied when the match_brackets setting is set to true.

bracket_contents_options

Controls certain options when the caret is in a bracket section. Only applied when the match_brackets setting is set to true.

brackets_foreground

Foreground color of the brackets when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

brackets_background

Background color of the brackets when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

brackets_options

Controls certain options when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

tags_foreground

Color of tags when the caret is next to a tag. Only used when the match_tags setting is set to true.

tags_options

Controls certain options when the caret is next to a tag. Only applied when the match_tags setting is set to true.

find_highlight

Background color of regions matching the current search.

find_highlight_foreground

Background color of regions matching the current search.

gutter

Background color of the gutter.

gutter_foreground

Foreground color of the gutter.

selection

Color of the selection regions.

selection_background

Background color of the selection regions.

selection_border

Color of the selection regions border.

inactive_selection

Color of inactive selections (inactive view).

guide

Color of the guides displayed to indicate nesting levels.

active_guide

Color of the guide lined up with the caret. Only applied if the indent_guide_options setting is set to draw_active.

stack_guide

Color of the current guide’s parent guide level. Only used if the indent_guide_options setting is set to draw_active.

highlight

Background color for regions added via sublime.add_regions() with the sublime.DRAW_OUTLINED flag added.

highlight_foreground

Foreground color for regions added via sublime.add_regions() with the sublime.DRAW_OUTLINED flag added.