syntect

Module html

Rendering highlighted code as HTML+CSS

Structs

ClassedHTMLGenerator

Output HTML for a line of code with <span> elements using class names

Enums

ClassStyle

IncludeBackground

Determines how background color attributes are generated

Functions

append_highlighted_html_for_styled_line

Like styled_line_to_highlighted_html but appends to a String for increased efficiency. In fact styled_line_to_highlighted_html is just a wrapper around this function.

css_for_themeDeprecated

css_for_theme_with_class_style

Create a complete CSS for a given theme. Can be used inline, or written to a CSS file.

highlighted_html_for_file

Convenience method that combines start_highlighted_html_snippet, styled_line_to_highlighted_html and HighlightFile from syntect::easy to create a full highlighted HTML snippet for a file.

highlighted_html_for_string

Convenience method that combines start_highlighted_html_snippet, styled_line_to_highlighted_html and HighlightLines from syntect::easy to create a full highlighted HTML snippet for a string (which can contain many lines).

line_tokens_to_classed_spans

Output HTML for a line of code with <span> elements specifying classes for each token. The span elements are nested like the scope stack and the scopes are mapped to classes based on the ClassStyle (see it’s docs).

start_highlighted_html_snippet

Returns a <pre style="...">\n tag with the correct background color for the given theme. This is for if you want to roll your own HTML output, you probably just want to use highlighted_html_for_string.

styled_line_to_highlighted_html

Output HTML for a line of code with <span> elements using inline style attributes to set the correct font attributes. The bg attribute determines if the spans will have the background-color attribute set. See the IncludeBackground enum’s docs.

tokens_to_classed_htmlDeprecated

tokens_to_classed_spansDeprecated

Preserved for compatibility, always use line_tokens_to_classed_spans and keep a ScopeStack between lines for correct highlighting that won’t sometimes crash.