Struct syntect::parsing::Scope

pub struct Scope {
    // some fields omitted
}

A hierarchy of atoms with semi-standardized names used to accord semantic information to a specific piece of text. Generally written with the atoms separated by dots. By convention atoms are all lowercase alphanumeric.

Example scopes: text.plain, punctuation.definition.string.begin.ruby, meta.function.parameters.rust

syntect uses an optimized format for storing these that allows super fast comparison and determining if one scope is a prefix of another. It also always takes 16 bytes of space. It accomplishes this by using a global repository to store string values and using bit-packed 16 bit numbers to represent and compare atoms. Like "atoms" or "symbols" in other languages. This means that while comparing and prefix are fast, extracting a string is relatively slower but ideally should be very rare.

Methods

impl Scope

fn new(s: &str) -> Result<Scope, ParseScopeError>

Parses a Scope from a series of atoms separated by . characters. Example: Scope::new("meta.rails.controller")

fn atom_at(self, index: usize) -> u16

Gets the atom number at a given index. I can't think of any reason you'd find this useful. It is used internally for turning a scope back into a string.

fn len(self) -> u32

return the number of atoms in the scope

fn build_string(self) -> String

returns a string representation of this scope, this requires locking a global repo and shouldn't be done frequently.

fn is_prefix_of(self, s: Scope) -> bool

Tests if this scope is a prefix of another scope. Note that the empty scope is always a prefix.

This operation uses bitwise operations and is very fast

Examples

use syntect::parsing::Scope;
assert!( Scope::new("string").unwrap()
        .is_prefix_of(Scope::new("string.quoted").unwrap()));
assert!( Scope::new("string.quoted").unwrap()
        .is_prefix_of(Scope::new("string.quoted").unwrap()));
assert!( Scope::new("").unwrap()
        .is_prefix_of(Scope::new("meta.rails.controller").unwrap()));
assert!(!Scope::new("source.php").unwrap()
        .is_prefix_of(Scope::new("source").unwrap()));
assert!(!Scope::new("source.php").unwrap()
        .is_prefix_of(Scope::new("source.ruby").unwrap()));
assert!(!Scope::new("meta.php").unwrap()
        .is_prefix_of(Scope::new("source.php").unwrap()));
assert!(!Scope::new("meta.php").unwrap()
        .is_prefix_of(Scope::new("source.php.wow").unwrap()));