syntect
Module util

Convenient helper functions for common use cases:

Printing to terminal

Iterating lines with \ns

Modifying ranges of highlighted output

Structs

LinesWithEndings

An iterator over the lines of a string, including the line endings.

Functions

as_24_bit_terminal_escaped

Formats the styled fragments using 24-bit color terminal escape codes. Meant for debugging and testing.

as_latex_escaped

Formats the styled fragments using LaTeX textcolor directive.

debug_print_ops

Print out the various push and pop operations in a vector with visual alignment to the line. Obviously for debugging.

modify_range

Modify part of a highlighted line using a style modifier, useful for highlighting sections of a line.

split_at

Split a highlighted line at a byte index in the line into a before and after component.